Set crunch/fastcgi to '2.x-dev' from 'dev-master'
I'll drop PHP5.x support and require PHP7+ some day (depending on free time and motivation). To avoid confusion I'll release it as "3.x" (although there were never an official 2.x-release). For you this means, that you should rely on the specific branch-alias for the 2.x-branch instead of `master`. Currently `master` and `2.x-dev` are identical.